+++ tests/testthat/test-bm_outline.R
@@ -2,10 +2,10 @@ test_that("`bm_outline()` works as expected", {
 	bm0 <- bm_bitmap(matrix(integer(), nrow = 4L, ncol = 0L))
 	expect_equal(bm0, bm_outline(bm0))
 
-	skip_if_not(cli::is_utf8_output())
 	skip_if_not_installed("farver")
 	skip_if_not_installed("magick")
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 
 	corn <- farming_crops_16x16()$corn$portrait