+++ tests/testthat/test-bm_pixmap.R
@@ -7,6 +7,7 @@ test_that("`bm_pixmap()` works as expected", {
 test_that("`as.array.bm_pixmap()`", {
 	skip_if_not_installed("png")
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 
 	tulip <- farming_crops_16x16()$tulip$portrait
@@ -49,8 +50,8 @@ test_that("`as_bm_pixmap()`", {
 
 test_that("`cbind()` and `rbind()`", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
-	skip_if_not(cli::is_utf8_output())
 	fc <- farming_crops_16x16()
 	avocado <- fc$avocado$portrait
 	tulip <- fc$tulip$portrait