+++ tests/testthat/test-bm_trim.R
@@ -4,6 +4,7 @@ capital_r <- font[[str2ucp("R")]]
 
 test_that("bm_trim()", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 	capital_r_trimmed <- bm_trim(capital_r, c(1, 1, 3, 0))
 	expect_equal(nrow(capital_r_trimmed), 12L)
@@ -31,8 +32,8 @@ test_that("bm_trim()", {
 
 test_that("bm_trim.bm_pixmap()", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
-	skip_if_not(cli::is_utf8_output())
 	crops <- farming_crops_16x16()
 	corn <- crops$corn$portrait
 	verify_output(
@@ -52,8 +53,8 @@ test_that("bm_trim.bm_pixmap()", {
 test_that("bm_trim.magick-image()", {
 	skip_if_not_installed("magick")
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
-	skip_if_not(cli::is_utf8_output())
 	crops <- farming_crops_16x16()
 	corn <- magick::image_read(crops$corn$portrait)
 	verify_output(
@@ -78,8 +79,8 @@ test_that("bm_trim.magick-image()", {
 test_that("bm_trim.nativeRaster()", {
 	skip_if_not_installed("farver")
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
-	skip_if_not(cli::is_utf8_output())
 	crops <- farming_crops_16x16()
 	corn <- as.raster(crops$corn$portrait, native = TRUE)
 	verify_output(
@@ -103,8 +104,8 @@ test_that("bm_trim.nativeRaster()", {
 
 test_that("bm_trim.raster()", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
-	skip_if_not(cli::is_utf8_output())
 	crops <- farming_crops_16x16()
 	corn <- as.raster(crops$corn$portrait)
 	verify_output(
@@ -128,6 +129,7 @@ test_that("bm_trim.raster()", {
 
 test_that("bm_shift()", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 	capital_r_shifted <- bm_shift(capital_r, bottom = 2L)
 	verify_output("txt/capital_r_shifted.txt", print(capital_r_shifted, px = c("-", "#")))
@@ -137,6 +139,7 @@ test_that("bm_shift()", {
 
 test_that("bm_shadow", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 	verify_output("txt/capital_r_shadow.txt", print(bm_shadow(capital_r), px = px_ascii))
 	verify_output("txt/capital_r_bold.txt", print(bm_bold(capital_r), px = px_ascii))
@@ -150,6 +153,7 @@ test_that("bm_shadow", {
 
 test_that("bm_rotate()", {
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 	verify_output("txt/capital_r_rotated90.txt", print(bm_rotate(capital_r, 90), px = px_ascii))
 	verify_output("txt/capital_r_rotated180.txt", print(bm_rotate(capital_r, 180), px = px_ascii))
@@ -157,6 +161,10 @@ test_that("bm_rotate()", {
 	verify_output(
 		"txt/capital_r_rotatedm90.txt",
 		print(bm_rotate(capital_r, 90, clockwise = FALSE), px = px_ascii)
+	)
+	expect_snapshot(
+		error = TRUE,
+		bm_rotate(capital_r, 90L, in_place = TRUE)
 	)
 })