+++ tests/testthat/test-fontr.R
@@ -2,6 +2,7 @@ test_that("`fontr` import works", {
 	skip_if_not_installed("fontr", "0.1")
 
 	skip_if_not_installed("withr")
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 
 	ch_R <- fontr::glyph_bitmap("R", family = "sans", face = "regular", pixel_size = 20)