+++ tests/testthat/test-pdftools.R
@@ -2,7 +2,7 @@ test_that("`pdftools::pdf_render_page()` import works", {
 	skip_if_not_installed("pdftools")
 
 	skip_if_not_installed("withr")
-	skip_if_not(cli::is_utf8_output())
+	withr::local_options(cli.unicode = TRUE)
 	withr::local_options(bm_options(default = TRUE))
 
 	filename <- tempfile(fileext = ".pdf")